+++ mitra_models/migrations/V0123__invoice__no_unique_payment_address.sql
@@ -1,0 +1,1 @@
+ALTER TABLE invoice DROP CONSTRAINT invoice_chain_id_payment_address_key;