+++ mitra_models/migrations/schema.sql
@@ -371,7 +371,6 @@ CREATE TABLE invoice (
     object_id VARCHAR(2000) UNIQUE,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    UNIQUE (chain_id, payment_address),
     CHECK (sender_id != recipient_id)
 );