+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Prevent changing invoice status to "Underpaid" if payout tx ID is set.
 - Prevent re-opening invoice if payout tx ID is not cleared.
 - Write payment amount to log when closing invoice.
+- Allow to create multiple invoices with the same payment address.
 
 ### Removed